Update agency.js
Updated the POST agency endpoint so now it store the agent id if the field is enabled for the insurer else it sets it to null rather than some place holder string value. If the insurer enables the agent id and no value is passed the endpoint returns an error. If we somehow get an insurer id which doesn't exist in the insurer's db we log warning (this should not ever happen but just sanity check)

Task ID #CU-8wcz9a