remove the limit from the owner set
it should not matter the limit since there should always be one, but a state can be reached where there are 2 and one is deleted. Thus allowing for 2 owners if the limit is left in.

The set of the new owner comes right after this, so it will have no effect on the behavior.