add ability to refresh token via an optional body parameter
req.body.refreshToken: is refreshToken a good name?

Moved some of the repeated logic to a central place "createNewToken" that handles values like ranValue1 and createdAt.

refreshThresholdSeconds: The the amount of seconds that need to have passed since creation, at 30 min with a 1 hour token, it will limit redis to 2 entries for the token on refresh.

Only allow refresh on POST/PUT (Application save). Should this be more broad?